alarms, alarms, alarms
added that alarms can make sound!
added that alarms can be triggered!
added that alarms can be snoozed!
added that alarms can be dismissed!
added that alarms also use light for wake up!
removed volume from alarms because it doesn't work that way!
fixed something unclear in stripHandling, now there is mid_up and mid_dn
(which are the same strips, but this makes splitting them for later
easier)
updated README!
some of this is still untested, most of it is though
good night